measure runtime: correctly labelled output

+++ 0x01-python_async_function/2-measure_runtime.py
@@ -7,7 +7,7 @@ import time
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-def measure_time(n: int, max_delay: int) -> int:
+def measure_time(n: int, max_delay: int) -> float:
     """
         measure_time: returns the average time taken to finish the process.
